requirements.txt
sentence-transformers
torch
transformers
huggingface_hub
accelerate
python-dotenv
pandas
tqdm
pillow
requests
sentencepiece
bitsandbytes
openai
anthropic
openrouter
qwen
google-genai
mistral_inference
qwen_vl_util

vllm --upgrade